feat(slides): try-it QR codes on demo slides + fullscreen play page
Each heavy-demo slide now shows a small Web + Lynx App QR pair (src/qrcodes.js),
both derived from location.origin at runtime so they work on any deploy domain:

- Web (also clickable) -> play.html?bundle=… , a fullscreen single-example
  player served from the deck's own origin (play.html + src/play.js).
- Lynx App -> the hosted <origin>/examples/<ex>/dist/<name>.lynx.bundle for the
  Lynx Explorer app to open natively (the .web.bundle's .lynx.bundle sibling).

QR is rendered to inline SVG (src/qr.js via qrcode-generator), always dark-on-
white so it scans in the dark theme. The <vl-demo> element is extracted into
src/demo.js and shared by the deck and the play page (registerVlDemo()).
play.html added as a Vite build input.

Per the chosen direction, the Web QR points at the deck's own play page for now;
buildUrls() is the single seam to later point it at the website Go component's
fullscreen deep-link (a separate PR to the website main branch).

Verified: 4 QR pairs render with correct web/native URLs, the web link opens the
play page and mounts the right bundle, and the QR SVGs are valid.
