+++ slides/package.json
@@ -11,7 +11,8 @@
   },
   "dependencies": {
     "@lynx-js/web-core": "^0.20.0",
-    "@lynx-js/web-elements": "^0.12.0"
+    "@lynx-js/web-elements": "^0.12.0",
+    "qrcode-generator": "^2.0.4"
   },
   "devDependencies": {
     "vite": "^5.4.10"
